moved boundMolecule fixture to conftest.py
